requirements.txt
Flask
openai
python-dotenv
pypdf
gunicorn  # 💡 خادم الإنتاج الجديد